removed warnings from type conversions.

+++ contrib/poly2tri/poly2tri/sweep/sweep_context.h
@@ -158,7 +158,7 @@ inline AdvancingFront* SweepContext::front()
 
 inline int SweepContext::point_count()
 {
-  return points_.size();
+  return (int)points_.size();
 }
 
 inline void SweepContext::set_head(Point* p1)
